third_party/blink/web_tests/wpt_internal/app-history/navigate-event/respondWith-on-synthetic-event.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws_dom("SecurityError", () => new AppHistoryNavigateEvent("navigate").respondWith(Promise.resolve()));
}, "AppHistoryNavigateEvent.respondWith throws if invoked on a synthetic event");
</script>
